Add files via upload
Observe and review properties that have experienced open code violations, sustained fire damages, or any other reasonings to why these types of properties are great opportunities for real estate Investors and also to see the many hardships real estate owners have to go through in order to come to the realization for them to sell their property because of the lack of financial support and stability in the light of those uncertain events. What we want to see here is compare 4 areas in  New York State and observe what real estate hardship each areas mostly experience based on the data I was able to retrieve and address it!